% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experts.R
\docType{data}
\name{experts}
\alias{experts}
\title{Data from QDA - Perfumes are evaluated by experts}
\format{
A dataframe with 288 rows and 16 variables:
\describe{
\item{Panelist}{dbl Year price was recorded}
\item{Session}{dbl Month price was recorded. Ranges from 1-12 for January - December}
\item{Rank}{dbl denoting monthly price for the 0-15/lb. size class}
\item{Product}{dbl denoting monthly price for the 15-20/lb. size class}
\item{Spicy-Wrapping}{dbl denoting monthly price for the 20-25/lb. size class}
}
}
\usage{
experts
}
\description{
Data from QDA - Perfumes are evaluated by experts
}
\keyword{datasets}
