% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tea.R
\docType{data}
\name{tea}
\alias{tea}
\title{Data tea: how do you drink your tea?}
\format{
A dataframe with 300 rows and 37 variables:
\describe{
\item{breakfast-type}{how do you drink your tea?}
\item{sex-frequency}{who are you?}
\item{exotic-no.effect.health}{how do you represent the product?}
\item{words}{how do you represent the product?}
}
}
\usage{
tea
}
\description{
Data tea: how do you drink your tea?
}
\keyword{datasets}
